# Conflicts:
#	h2/src/main/org/h2/command/Command.java
#	h2/src/main/org/h2/command/CommandRemote.java
#	h2/src/main/org/h2/command/Parser.java
#	h2/src/main/org/h2/command/Prepared.java
#	h2/src/main/org/h2/command/ddl/AlterTableAlterColumn.java
#	h2/src/main/org/h2/command/ddl/CreateDomain.java
#	h2/src/main/org/h2/command/ddl/CreateIndex.java
#	h2/src/main/org/h2/command/ddl/CreateTable.java
#	h2/src/main/org/h2/command/dml/BackupCommand.java
#	h2/src/main/org/h2/command/dml/Call.java
#	h2/src/main/org/h2/command/dml/Delete.java
#	h2/src/main/org/h2/command/dml/Insert.java
#	h2/src/main/org/h2/command/dml/Merge.java
#	h2/src/main/org/h2/command/dml/SetTypes.java
#	h2/src/main/org/h2/command/dml/Update.java
#	h2/src/main/org/h2/command/query/Optimizer.java
#	h2/src/main/org/h2/command/query/Query.java
#	h2/src/main/org/h2/command/query/Select.java
#	h2/src/main/org/h2/command/query/SelectUnion.java
#	h2/src/main/org/h2/constraint/Constraint.java
#	h2/src/main/org/h2/constraint/ConstraintCheck.java
#	h2/src/main/org/h2/constraint/ConstraintReferential.java
#	h2/src/main/org/h2/engine/ConnectionInfo.java
#	h2/src/main/org/h2/engine/Database.java
#	h2/src/main/org/h2/engine/DbObject.java
#	h2/src/main/org/h2/engine/Engine.java
#	h2/src/main/org/h2/engine/Right.java
#	h2/src/main/org/h2/engine/Role.java
#	h2/src/main/org/h2/engine/Session.java
#	h2/src/main/org/h2/engine/SessionRemote.java
#	h2/src/main/org/h2/engine/User.java
#	h2/src/main/org/h2/expression/Expression.java
#	h2/src/main/org/h2/expression/ExpressionColumn.java
#	h2/src/main/org/h2/expression/ValueExpression.java
#	h2/src/main/org/h2/expression/Variable.java
#	h2/src/main/org/h2/expression/Wildcard.java
#	h2/src/main/org/h2/expression/aggregate/Aggregate.java
#	h2/src/main/org/h2/expression/analysis/DataAnalysisOperation.java
#	h2/src/main/org/h2/expression/condition/Comparison.java
#	h2/src/main/org/h2/expression/condition/ConditionAndOr.java
#	h2/src/main/org/h2/expression/condition/ConditionIn.java
#	h2/src/main/org/h2/expression/condition/ConditionInQuery.java
#	h2/src/main/org/h2/expression/function/DateTimeFunctions.java
#	h2/src/main/org/h2/expression/function/Function.java
#	h2/src/main/org/h2/index/BaseIndex.java
#	h2/src/main/org/h2/index/Index.java
#	h2/src/main/org/h2/index/ViewIndex.java
#	h2/src/main/org/h2/jdbc/JdbcPreparedStatement.java
#	h2/src/main/org/h2/jdbc/JdbcStatement.java
#	h2/src/main/org/h2/mode/FunctionInfo.java
#	h2/src/main/org/h2/mvstore/Cursor.java
#	h2/src/main/org/h2/mvstore/CursorPos.java
#	h2/src/main/org/h2/mvstore/MVStore.java
#	h2/src/main/org/h2/mvstore/Page.java
#	h2/src/main/org/h2/mvstore/db/MVPrimaryIndex.java
#	h2/src/main/org/h2/mvstore/db/MVSecondaryIndex.java
#	h2/src/main/org/h2/mvstore/db/MVTableEngine.java
#	h2/src/main/org/h2/pagestore/PageLog.java
#	h2/src/main/org/h2/pagestore/PageStore.java
#	h2/src/main/org/h2/pagestore/db/PageBtreeIndex.java
#	h2/src/main/org/h2/pagestore/db/PageBtreeLeaf.java
#	h2/src/main/org/h2/pagestore/db/PageBtreeNode.java
#	h2/src/main/org/h2/pagestore/db/ScanIndex.java
#	h2/src/main/org/h2/result/ResultRemote.java
#	h2/src/main/org/h2/schema/Schema.java
#	h2/src/main/org/h2/schema/TriggerObject.java
#	h2/src/main/org/h2/server/TcpServerThread.java
#	h2/src/main/org/h2/server/pg/PgServer.java
#	h2/src/main/org/h2/table/Column.java
#	h2/src/main/org/h2/table/FunctionTable.java
#	h2/src/main/org/h2/table/RegularTable.java
#	h2/src/main/org/h2/table/Table.java
#	h2/src/main/org/h2/table/TableFilter.java
#	h2/src/main/org/h2/table/TableView.java
#	h2/src/main/org/h2/value/DataType.java
#	h2/src/main/org/h2/value/Transfer.java
#	h2/src/main/org/h2/value/Value.java
#	h2/src/main/org/h2/value/ValueBigint.java
#	h2/src/main/org/h2/value/ValueInteger.java
#	h2/src/main/org/h2/value/ValueLob.java
#	h2/src/main/org/h2/value/ValueString.java

+++ h2/src/test/org/h2/test/db/TestAlter.java
@@ -10,7 +10,13 @@ import java.sql.DatabaseMetaData;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.util.Collection;
+
 import org.h2.api.ErrorCode;
+import org.h2.engine.Session;
+import org.h2.engine.SessionLocal;
+import org.h2.jdbc.JdbcConnection;
+import org.h2.schema.Sequence;
 import org.h2.test.TestBase;
 import org.h2.test.TestDb;
 
@@ -28,7 +34,7 @@ public class TestAlter extends TestDb {
      * @param a ignored
      */
     public static void main(String... a) throws Exception {
-        TestBase.createCaller().init().test();
+        TestBase.createCaller().init().testFromMain();
     }
 
     @Override
@@ -123,22 +129,25 @@ public class TestAlter extends TestDb {
     }
 
     private void testAlterTableDropIdentityColumn() throws SQLException {
+        Session iface = ((JdbcConnection) stat.getConnection()).getSession();
+        if (!(iface instanceof SessionLocal)) {
+            return;
+        }
+        Collection<Sequence> allSequences = ((SessionLocal) iface).getDatabase().getMainSchema().getAllSequences();
         stat.execute("create table test(id int auto_increment, name varchar)");
         stat.execute("alter table test drop column id");
-        ResultSet rs = stat.executeQuery("select * from INFORMATION_SCHEMA.SEQUENCES");
-        assertFalse(rs.next());
+        assertEquals(0, allSequences.size());
         stat.execute("drop table test");
 
         stat.execute("create table test(id int auto_increment, name varchar)");
         stat.execute("alter table test drop column name");
-        rs = stat.executeQuery("select * from INFORMATION_SCHEMA.SEQUENCES");
-        assertTrue(rs.next());
+        assertEquals(1, allSequences.size());
         stat.execute("drop table test");
     }
 
     private void testAlterTableAddColumnIdentity() throws SQLException {
         stat.execute("create table t(x varchar)");
-        stat.execute("alter table t add id bigint identity(5, 5) not null");
+        stat.execute("alter table t add id bigint identity(5, 5) default on null");
         stat.execute("insert into t values (null, null)");
         stat.execute("insert into t values (null, null)");
         ResultSet rs = stat.executeQuery("select id from t order by id");
